feat: Add phone and category to User type
This change introduces `phone` and `category` properties to the `User` interface. This is necessary to support displaying contact information and categorizing users within the application, particularly for marketplace and contact list functionalities. The `BusinessProfile` interface has also been removed and its relevant fields consolidated into the `User` interface to simplify the data structure.